+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Added `Suppliers.supplyTrue` to always supply `true`.
 - Added `Suppliers.supplyFalse` to always supply `false`.
 - Added `Suppliers.supplyNull` to always return `null`.
+- Added `Suppliers.supplyNull(Runnable)` which returns a supplier that runs the runnable and then returns `null`.
+- Added `Suppliers.compose(Runnable, Supplier)` which returns a supplier that runs the runnable and the returns the supplied value by the given supplier.
 - Added `IntPredicates.alwaysTrue` to return a predicate that is always false.
 - Added `IntPredicates.alwaysFalse` to return a predicate that is always false.
 - Added `IntPredicates.rejectAll` to return a predicate that is always false (alias to `alwaysFalse`).